requirements.txt
# === Core ===
numpy>=1.24
pandas>=2.0

# === Deep Learning ===
torch>=2.8
torchvision>=0.23
torchmetrics>=1.8

# === Kaggle datasets ===
kaggle>=1.7
kagglehub>=0.3

# === Utilities ===
requests>=2.32
tqdm>=4.67
Pillow>=11.0

# === Config management ===
hydra-core>=1.3
omegaconf>=2.3
PyYAML>=6.0

# === Optional / text utilities ===
python-slugify>=8.0
text-unidecode>=1.3
bleach>=6.0
